RSUSDEM Update v2022
This update to RSUSDEM and the Functions and Classes brings the version of this module up to date with the most developed succesful run version of the calibration module run on a test case of Haringey in 2022.

Specifically this now contains the correct loading commands for a native R run.

Next update to include plotting and processing function update which had version issues/bugs that needed resolving.

Requires the Haringey Test data to perform a full run.